add  docs e altera do double(Double)

+++ backend/src/main/java/com/juliuseconomyapi/backend/model/Finance.java
@@ -27,7 +27,7 @@ public class Finance {
 
    
     @Column(nullable = false)
-    private BigDecimal value;
+    private double value;
 
     @Column(length = 50)
     private String category;
